requirements.txt
Flask
python-docx
nltk
spacy
scikit-learn
Werkzeug
gunicorn
flask-cors
pdfplumber